requirements.txt
django
maxapi[webhook]
python-dotenv
requests
psycopg2-binary
zulip
